app/requirements.txt
Django==4.2.2
django-bootstrap-v5==1.0.11
django-cors-headers==4.1.0
django-resized==1.0.2
djangorestframework==3.14.0
djangorestframework-camel-case==1.4.2
google-api-core==2.11.1
google-api-python-client==2.91.0
google-auth==2.21.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.1
oauthlib==3.2.2
psycopg2-binary==2.9.6
gunicorn==21.2.0
Pillow==9.5.0